# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.AnomalyItem:
    properties:
      anomaly_score:
        type: number
      category:
        type: string
      price:
        type: number
      product_id:
        type: string
      product_name:
        type: string
      unit:
        type: string
    type: object
...
  entities.ClassifyProductRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  entities.ClassifyProductResponse:
    properties:
      category:
        type: string
      confidence:
        type: number
      original_name:
        type: string
      standard_id:
        type: integer
      standard_name:
        type: string
    type: object
  entities.CreateProductRequest:
    properties:
      category:
        maxLength: 100
        minLength: 1
        type: string
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      price:
        type: number
      unit:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - category
    - name
    - price
    - unit
    type: object
...
  entities.DuplicatePair:
    properties:
      product_id_1:
        type: string
      product_id_2:
        type: string
      product_name_1:
        type: string
      product_name_2:
        type: string
      similarity_score:
        type: number
    type: object
...
  entities.Product:
    properties:
      active:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      provider_id:
        type: string
      unit:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.UpdateProductRequest:
    properties:
      category:
        maxLength: 100
        minLength: 1
        type: string
      description:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      price:
        type: number
      unit:
        maxLength: 50
        minLength: 1
        type: string
    type: object
...
  responses.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: soporte@visionprice.app
    name: VisionPrice Support
  description: API backend para el módulo de Proveedores de VisionPrice
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VisionPrice Proveedores API
  version: "1.0"
paths:
...
  /api/v1/ml/products/anomalies:
    get:
      description: Utiliza Isolation Forest comparando con el mercado global para
        detectar precios anómalos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.AnomalyItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Detectar anomalías en precios
      tags:
      - Machine Learning
  /api/v1/ml/products/classify:
    post:
      consumes:
      - application/json
      description: Utiliza TF-IDF y Similitud Coseno para mapear un nombre de producto
        al catálogo maestro
      parameters:
      - description: Nombre del producto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.ClassifyProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.ClassifyProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Clasificar producto automáticamente
      tags:
      - Machine Learning
  /api/v1/ml/products/duplicates:
    get:
      description: Encuentra posibles productos duplicados en el catálogo del proveedor
        logueado usando TF-IDF
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.DuplicatePair'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Detectar productos duplicados
      tags:
      - Machine Learning
  /api/v1/products:
    get:
      description: Retorna todos los productos activos del proveedor autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Product'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Listar productos activos
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Registra un nuevo material/producto para el proveedor autenticado
      parameters:
      - description: Datos del producto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Crear un nuevo producto
      tags:
      - Products
  /api/v1/products/{id}:
    delete:
      description: Desactiva un producto del proveedor autenticado (soft delete)
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Eliminar producto (baja lógica)
      tags:
      - Products
    get:
      description: Retorna un producto específico del proveedor autenticado
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Product'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Obtener producto por ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un producto existente del proveedor autenticado
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Actualizar producto
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'Ingrese el token JWT con el prefijo Bearer. Ejemplo: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"